+++ server/pigallery.js
@@ -8,6 +8,7 @@ const express = require('express');
 const session = require('express-session');
 const FileStore = require('session-file-store')(session);
 const nedb = require('nedb-promises');
+const MongoClient = require("mongodb").MongoClient;
 const api = require('./api.js');
 const build = require('./build.js');
 const watcher = require('./watcher.js');
@@ -81,10 +82,6 @@ async function main() {
     log.error('Configuration missing, exiting...');
     process.exit(0);
   }
-
-  // create image db backup
-  if (!fs.existsSync(config.server.db)) log.warn('Image cache not found:', config.server.db);
-  else await createBackup(config.server.db);
 
   // initialize esbuild bundler
   await build.compile();
@@ -189,18 +186,59 @@ async function main() {
   }
 
   // load image cache
-  db = await nedb.create({ filename: config.server.db, inMemoryOnly: false, timestampData: true, autoload: false });
-  await db.ensureIndex({ fieldName: 'image', unique: true, sparse: true });
-  await db.ensureIndex({ fieldName: 'processed', unique: false, sparse: false });
-  await db.load();
-  const records = await db.count({});
-  log.state('Image DB loaded:', config.server.db, 'records:', records);
+  log.info('Database engine:', config.server.db);
+  if (config.server.db === 'nedb') {
+    // create image db backup
+    if (!fs.existsSync(config.server.nedb)) log.warn('Image cache not found:', config.server.nedb);
+    else await createBackup(config.server.nedb);
 
-  const shares = await db.find({ images: { $exists: true } });
-  for (const share of shares) {
-    log.state('Shares:', share.name, 'creator:', share.creator, 'key:', share.share, 'images:', share.images.length);
+    db = await nedb.create({ filename: config.server.nedb, inMemoryOnly: false, timestampData: true, autoload: false });
+    await db.ensureIndex({ fieldName: 'image', unique: true, sparse: true });
+    await db.ensureIndex({ fieldName: 'processed', unique: false, sparse: false });
+    await db.load();
+    const records = await db.count({});
+    log.state('Image DB loaded:', config.server.nedb, 'records:', records);
+
+    const shares = await db.find({ images: { $exists: true } });
+    for (const share of shares) {
+      log.state('Shares:', share.name, 'creator:', share.creator, 'key:', share.share, 'images:', share.images.length);
+    }
   }
+  if (config.server.db === 'mongodb') {
+    // connect to mongodb instance
+    const client = new MongoClient(config.server.mongoURI, config.server.mongoOptions);
+    log.info('MongoDB Client:', client['s'].options.hosts[0]);
+    log.info('MongoDB Driver:', client['s'].options.metadata.driver);
+    client.on('commandFailed', (event) => log.warn('MongoDB command:', event.commandName, event.failure));
+    client.on('serverOpening', () => log.state('MongoDB server connection opening'));
+    client.on('serverClosed', () => log.state('MongoDB server connection closed'));
+    try {
+      await client.connect();
+    } catch (err) {
+      log.error('MongoDB:', err.errmsg || err);
+    }
+    const state = client['topology'].s.state;
+    log.state('MongoDB State:', state);
+    if (state === 'closed') process.exit(1);
+    // init mongodb database
+    const database = client.db(config.server.mongoDB);
+    const images = database.collection('images');
+    const shares = database.collection('shares');
+    await images.createIndex({ 'image': 1 });
+    await images.createIndex({ 'processed': 1 });
 
+    // mongodb state before migration
+    log.info('Image DB loaded:', await images.countDocuments());
+    log.info('Shares loaded:', await shares.countDocuments());
+
+    // TBD
+    log.warn('MongoDB is not yet supported');
+    process.exit(1);
+  }
+  if (!db) {
+    log.error('Database engine not recognized:', config.server.db);
+    process.exit(1);
+  }
   // initialize api calls
   api.init(app, config, db);
 }