+++ CHANGELOG.md
@@ -9,7 +9,11 @@ Repository: **<git+https://github.com/vladmandic/pigallery.git>**
 
 ## Changelog
 
-### **HEAD -> master** 2021/08/05 mandic00@live.com
+### **HEAD -> master** 2021/08/22 mandic00@live.com
+
+- enable large database
+
+### **origin/master** 2021/08/05 mandic00@live.com
 
 - enable uniform support in latest tfjs
 - merge pull request #8 from trissiethehusky/master